Modified remaining shaders and renderers in the OXR version to read the palette from the core engine using an uniform buffer instead of a texture. Modified location of the "time" variable for turbulent surfaces at the beginning of the push constants buffer - with this, there is no longer a need for a specialized turbulent_rotated fragment shader. Removed "palette" as a texture from PerImage; palette is now what used to be paletteAsBuffer from the previous commit. Removed no longer needed bufferAndImageLayout from the renderer in the OXR version.

+++ OpenXR-SDK/src/SlipNFrag-XR/app/src/main/shaders/colored.vert
@@ -16,7 +16,7 @@ layout(set = 0, binding = 0) uniform SceneMatrices
 
 layout(location = 0) in vec3 vertexPosition;
 layout(location = 1) in float vertexColor;
-layout(location = 0) out float fragmentColor;
+layout(location = 0) out int fragmentColor;
 
 out gl_PerVertex
 {
@@ -26,5 +26,5 @@ out gl_PerVertex
 void main(void)
 {
 	gl_Position = projectionMatrix[gl_ViewID_OVR] * (viewMatrix[gl_ViewID_OVR] * (vertexTransform * vec4(vertexPosition, 1)));
-	fragmentColor = vertexColor;
+	fragmentColor = int(vertexColor);
 }
